refactor: settings-about
